add websockets
Took 18 minutes
