+++ ServerHttp/pom.xml
@@ -87,6 +87,12 @@
             <version>${ktor.version}</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>io.ktor</groupId>
+            <artifactId>ktor-websockets</artifactId>
+            <version>${ktor.version}</version>
+            <scope>compile</scope>
+        </dependency>
         <!--Ktor-->
 
         <dependency>
